% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FFP_to_raster.R
\name{FFP_to_raster}
\alias{FFP_to_raster}
\title{Converts Footprint values to a raster}
\usage{
FFP_to_raster(
  FFP,
  site_long,
  site_lat,
  project_raster = FALSE,
  project_crs_epsg = 3006
)
}
\arguments{
\item{FFP}{output of the FFP model, the footprint climatology model}

\item{site_long}{site longitude}

\item{site_lat}{site latitude}

\item{project_raster}{boolean, default is FALSE. Set to TRUE to project the raster into a different CRS. Make sure to set the output EPSG correctly}

\item{project_crs_epsg}{EPSG code of the output projection of the FFP raster}
}
\value{
a raster in CRS: EPSG 4326, with the footprint values
}
\description{
Converts Footprint values to a raster
}
