% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{site_coords_example}
\alias{site_coords_example}
\title{Example Site Coordinates for Halmyran}
\format{
A data frame with 1 row and 2 variables:
\describe{
  \item{long}{Longitude in decimal degrees (WGS84)}
  \item{lat}{Latitude in decimal degrees (WGS84)}
}
}
\source{
Halmyran peatland research site, Sweden
}
\usage{
site_coords_example
}
\description{
Location of the Halmyran eddy covariance flux tower in Sweden.
}
\examples{
data(site_coords_example)
site_coords_example
}
\keyword{datasets}
