% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xr_yr_2_lat_long.R
\name{xr_yr_2_lat_long}
\alias{xr_yr_2_lat_long}
\title{Convertion of xr and yr from the FFP model to long and lat}
\usage{
xr_yr_2_lat_long(site_lat, site_long, xr, yr, R = 6371007.181)
}
\arguments{
\item{site_lat}{site latitude}

\item{site_long}{site longitude}

\item{xr}{x-array _A list_ for contour line of r, r is the footprint percentage}

\item{yr}{y-array _A list_ for contour line of r}

\item{R}{the radius of the idealized sphere representing the Earth, in meter. Default value according to IUGG is 6371007.1810 m}
}
\value{
a named list of latitudes and longitudes
}
\description{
Convertion of xr and yr from the FFP model to long and lat
}
